[Chore] #110 -SearhMajor 디렉토리 생성, SerachMajorVC 및 SearchTableVIewCell  이동
