+++ GNUting/GNUting/Presentation/SearchMajor/SearchTableViewCell.swift
@@ -5,17 +5,27 @@
 //  Created by 원동진 on 3/19/24.
 //
 
+// MARK: - 학과 검색 VC TableView Cell
 
 import UIKit
 
-class MajorSearchTableViewCell: UITableViewCell {
-    static let identi = "MajorSearchTableViewCellid"
+class SearchTableViewCell: UITableViewCell {
+    // MARK: - Properties
+    
+    static let identi = "SearchTableViewCellid"
+    
+    // MARK: - SubViews
+    
     private lazy var majorLabel: UILabel = {
         let label = UILabel()
         label.font = Pretendard.bold(size: 16)
         label.textAlignment = .left
+        
         return label
     }()
+    
+    // MARK: - Init
+    
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         configure()
@@ -24,10 +34,12 @@ class MajorSearchTableViewCell: UITableViewCell {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
 }
 
-extension MajorSearchTableViewCell {
+extension SearchTableViewCell {
+    
+    // MARK: - Layout Helpers
+    
     private func configure() {
         contentView.addSubview(majorLabel)
         majorLabel.snp.makeConstraints { make in
@@ -36,8 +48,9 @@ extension MajorSearchTableViewCell {
             make.bottom.equalToSuperview().offset(-15)
         }
     }
-}
-extension MajorSearchTableViewCell {
+    
+    // MARK: - SetView
+    
     func setCell(model : SearchMajorModelResult){
         majorLabel.text = model.name
     }
